* fix: don't add exception when ignored conditions and notify (encountered   conditions must *all* be set in options for exception to be added !) * add issuer_not_trusted bypass condition (obviously related to self-signed) * add bypass conditions to notificationBox when adding exception * fix: sce.UIOptions.onLoad() was not called upon load event
NOTE: we are relying on nsIRecentBadCertsService, but it may not always tell
about a bad cert. For ex. when a revoked cert provides OCSP server
information. ...probably because the OCSP server is queried after
onSecurityChange.
We could still get the SSLStatus and the cert (d9e396f5), but might not be
able to tell, for certain, that the cert is bad. Because we might not know all
errors we'd have to check... So, for now, let's just not support OCSP.